started to refactor button.
