+++ src/components/button/Button.tsx
@@ -3,10 +3,11 @@ import React, {
   ForwardRefRenderFunction,
   HTMLAttributes,
 } from "react";
-import styled from "@emotion/styled";
-import { COLORS } from "../constants";
 
-type Variant = "primary" | "outline";
+import { COLORS } from "./constants";
+import { BaseButton } from "./Button.styled";
+
+type Variant = "primary" | "outline" | "base";
 export type Props = {
   variant?: Variant;
 } & HTMLAttributes<HTMLButtonElement>;
@@ -36,17 +37,25 @@ const STYLES: {
     "--hoverBackgroundColor": COLORS.primary,
     "--hoverTextColor": COLORS.white,
   },
+  base: {
+    "--textColor": "",
+    "--bgColor": "",
+    "--borderColor": "",
+    "--hoverBackgroundColor": "",
+    "--hoverTextColor": ""
+  }
 };
 
 const ButtonComponent: ForwardRefRenderFunction<HTMLButtonElement, Props> = (
   { children, variant = "primary", ...delegated },
   externalRef
 ) => {
-  if (!["primary", "outline"].includes(variant)) {
+  if (!["primary", "outline", "base"].includes(variant)) {
     throw new Error(
       `Button should be passed a valid variant prop, you passed ${variant}. If you don't pass a variant prop, Button will default to primary.`
     );
   }
+
   const styles = STYLES[variant];
   return (
     <BaseButton style={styles} ref={externalRef} {...delegated}>
@@ -54,22 +63,8 @@ const ButtonComponent: ForwardRefRenderFunction<HTMLButtonElement, Props> = (
     </BaseButton>
   );
 };
+
 const Button = forwardRef(ButtonComponent);
 Button.displayName = "Button";
+
 export default Button;
-
-const BaseButton = styled.button`
-  padding: 15px;
-  border-radius: 4px;
-  border: 1px solid var(--borderColor);
-  color: var(--textColor);
-  background-color: var(--bgColor);
-  transition-property: background-color, border-color, color, fill, stroke;
-  transition-timing-function: cubic-bezier(0.4, 0, 0.2, 1);
-  transition-duration: 150ms;
-
-  &:hover {
-    color: var(--hoverTextColor);
-    background-color: var(--hoverBackgroundColor);
-  }
-`;
