+++ src/components/button/index.ts
@@ -1,0 +1,1 @@
+export { default as Button, Props as ButtonProps } from "./Button";
