+++ src/components/index.ts
@@ -1,1 +1,1 @@
-export { default as Button, Props as ButtonProps } from "./Button";
+export { Button, ButtonProps } from "./button";
